add githubv4IsNodeNotFoundError
